+++ gamdpy/interactions/potential_functions/__init__.py
@@ -2,6 +2,7 @@
 from .LJ_12_6 import LJ_12_6
 from .LJ_12_6_sigma_epsilon import LJ_12_6_sigma_epsilon 
 from .SAAP import SAAP
+from .universal_zbl_potential import universal_zbl_potential
 from .harmonic_repulsion import harmonic_repulsion
 from .hertzian import hertzian
 from .LJ_SF import LJ_SF